Convert timeslice levels to lowercase
